# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Auth Login Protect API",
    "version": "1.0.0",
    "description": "Authentication and protected API routes using Supabase Auth"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Create a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },

    "/auth/login": {
      "post": {
        "summary": "Login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },

    "/auth/logout": {
      "post": {
        "summary": "Logout authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Logout successful"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },

    "/public/info": {
      "get": {
        "summary": "Get public information",
        "responses": {
          "200": {
            "description": "Public information"
          }
        }
      }
    },

    "/protected/profile": {
      "get": {
        "summary": "Get authenticated user profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated user profile"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },

    "/protected/dashboard": {
      "get": {
        "summary": "Get protected dashboard",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Protected dashboard"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  }
}